+++ wily/operators/raw.py
@@ -16,14 +16,18 @@ class RawMetricsOperator(BaseOperator):
     name = "raw"
     defaults = {"exclude": None, "ignore": None, "summary": False}
     metrics = (
-        Metric("loc", "Lines of Code", int, MetricType.Informational),
-        Metric("lloc", "L Lines of Code", int, MetricType.AimLow),
-        Metric("sloc", "S Lines of Code", int, MetricType.AimLow),
-        Metric("comments", "Multi-line comments", int, MetricType.AimHigh),
-        Metric("multi", "Multi lines", int, MetricType.Informational),
-        Metric("blank", "blank lines", int, MetricType.Informational),
+        Metric("loc", "Lines of Code", int, MetricType.Informational, sum),
+        Metric("lloc", "L Lines of Code", int, MetricType.AimLow, sum),
+        Metric("sloc", "S Lines of Code", int, MetricType.AimLow, sum),
+        Metric("comments", "Multi-line comments", int, MetricType.AimHigh, sum),
+        Metric("multi", "Multi lines", int, MetricType.Informational, sum),
+        Metric("blank", "blank lines", int, MetricType.Informational, sum),
         Metric(
-            "single_comments", "Single comment lines", int, MetricType.Informational
+            "single_comments",
+            "Single comment lines",
+            int,
+            MetricType.Informational,
+            sum,
         ),
     )
     default_metric_index = 0  # LOC